requirements.txt
twitchio
openai
python-dotenv